+++ tailwind.config.js
@@ -1,12 +1,5 @@
 module.exports = {
     darkMode: "class",
-    content: [
-        "./resources/**/*.blade.php",
-        "./resources/**/*.js",
-        "./resources/**/*.vue",
-    ],
-    theme: {
-        extend: {},
-    },
+    content: ["./resources/**/*.blade.php", "./resources/**/*.js", "./resources/**/*.vue"],
     plugins: [],
 };